requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.3
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
pytest==7.4.4
black==23.12.1
isort==5.13.2
flake8==7.0.0
python-amazon-paapi==5.0.1 